Debugging
After playing the game for a while, I noticed that when the car crashes into an obstacle right when the math quiz pops up, the code still runs after crashing. I tried to get rid of the bug by repeating the loop to check collisions first inside the mathquiz loop. Took about 3.5 hours to figure out how to fix it.